Moved PyMimeData to glue.utils.qt.mime
