+++ glue/core/data_factories/tests/test_data_factories.py
@@ -7,13 +7,12 @@ import numpy as np
 from mock import MagicMock
 from numpy.testing import assert_allclose, assert_array_equal
 
-from glue.core.tests.util import make_file
 from glue.core.component import CategoricalComponent
 from glue.core.data import Data
 from glue.core import data_factories as df
 from glue.config import data_factory
 from glue.tests.helpers import (requires_astropy, requires_astropy_ge_03,
-                                requires_pil_or_skimage)
+                                requires_pil_or_skimage, make_file)
 
 
 def test_load_data_auto_assigns_label():
