+++ glue/core/data_factories/tests/test_excel.py
@@ -4,9 +4,8 @@ import os
 
 from numpy.testing import assert_array_equal
 
-from glue.core.tests.util import make_file
 from glue.core import data_factories as df
-from glue.tests.helpers import requires_xlrd
+from glue.tests.helpers import requires_xlrd, make_file
 
 
 DATA = os.path.join(os.path.dirname(__file__), 'data')
