+++ glue/core/tests/test_coordinate_links.py
@@ -7,8 +7,7 @@ from glue.tests.helpers import requires_astropy, ASTROPY_INSTALLED
 from .. import Data, DataCollection
 from ..coordinates import coordinates_from_header
 from ..link_helpers import LinkSame
-from .util import make_file
-
+from glue.tests.helpers import make_file
 
 if ASTROPY_INSTALLED:
     from astropy.io import fits
