+++ glue/core/tests/util.py
@@ -1,37 +1,12 @@
 from __future__ import absolute_import, division, print_function
 
-import os
-import zlib
-import tempfile
 from contextlib import contextmanager
 
 from mock import MagicMock
 
-from glue.core.application_base import Application
 from glue import core
-
-
-@contextmanager
-def make_file(contents, suffix, decompress=False):
-    """Context manager to write data to a temporary file,
-    and delete on exit
-
-    :param contents: Data to write. string
-    :param suffix: File suffix. string
-    """
-    if decompress:
-        contents = zlib.decompress(contents)
-
-    try:
-        _, fname = tempfile.mkstemp(suffix=suffix)
-        with open(fname, 'wb') as outfile:
-            outfile.write(contents)
-        yield fname
-    finally:
-        try:
-            os.unlink(fname)
-        except WindowsError:  # on Windows the unlink can fail
-            pass
+from glue.core.application_base import Application
+from glue.tests.helpers import make_file
 
 
 @contextmanager
